move game info functions from game_controller to mahjong/game
